Allow author to be None
During a sync today I ran into an issue where the following error
occurred: `Errors during transform: [author_id: None does not match {'type': 'integer'}`

The data element had the author key missing, so it couldn't dig into the
sub resource to determine the `id` key.

This change allows the author for a release to be `nil` in case the key
is missing from the response.