Allow start_date to override state timestamp
Allows the start_date to be set to a more recent time than previous runs
and for this to clamp how far back the tap will search Gitlab for even
if the state wants to check further back than the start_date.

This allows users to ignore the state trying to search back to the first
'created' status job such as when a pipeline has a manual job that is
also blocking (allow_failure: false) and jobs that come after it.

Closes https://gitlab.com/meltano/tap-gitlab/-/issues/31